+++ packages/sdk-install-modal-web/CHANGELOG.md
@@ -6,6 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.32.1]
+### Added
+- chore: updates assets ([#1220](https://github.com/MetaMask/metamask-sdk/pull/1220))
+
+## [0.32.0]
+### Added
+- feat: setup stencil development environment with component preview ([#1196](https://github.com/MetaMask/metamask-sdk/pull/1196))
+
+### Fixed
+- fix: user rejection bug ([#1202](https://github.com/MetaMask/metamask-sdk/pull/1202))
+
+## [0.31.5]
+### Added
+- chore: add analytics to install modal ([#1189](https://github.com/MetaMask/metamask-sdk/pull/1189))
+
+### Fixed
+- fix: correct event handler binding in modal components ([#1191](https://github.com/MetaMask/metamask-sdk/pull/1191))
+- fix: remove console log when render install modal ([#1190](https://github.com/MetaMask/metamask-sdk/pull/1190))
+
+## [0.31.2]
+### Fixed
+- Set initial modal tab based on preferDesktop option ([#1158](https://github.com/MetaMask/metamask-sdk/pull/1158))
+
 ## [0.31.1]
 ### Changed
 - fix: Tell webpack about dynamic import + fixed polyfills ([#1151](https://github.com/MetaMask/metamask-sdk/pull/1151))
@@ -169,7 +192,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update GitHub actions workflows ([#102](https://github.com/MetaMask/metamask-sdk/pull/102))
 - [FEAT] Yarn v3 migration ([#100](https://github.com/MetaMask/metamask-sdk/pull/100))
 
-[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.31.1...HEAD
+[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.32.1...HEAD
+[0.32.1]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.32.0...@metamask/sdk-install-modal-web@0.32.1
+[0.32.0]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.31.5...@metamask/sdk-install-modal-web@0.32.0
+[0.31.5]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.31.2...@metamask/sdk-install-modal-web@0.31.5
+[0.31.2]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.31.1...@metamask/sdk-install-modal-web@0.31.2
 [0.31.1]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.31.0...@metamask/sdk-install-modal-web@0.31.1
 [0.31.0]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.30.2...@metamask/sdk-install-modal-web@0.31.0
 [0.30.2]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-install-modal-web@0.30.0...@metamask/sdk-install-modal-web@0.30.2